Bug 1646708 - Reset sweepMarkTaskStarted after joining the sweep task r=sfink
The problem here is that if we don't reset sweepMarkTaskStarted then we won't start the task next time there is work for it to do. This was missing from endSweepingSweepGroup but I've refectored so this happens in joinSweepMarkTask now.

Differential Revision: https://phabricator.services.mozilla.com/D80441

UltraBlame original commit: 14c68ad90aaa1d41e0179accee9d244b62997aec