Bug 1647702 - Broaden the race supression for compacting race between updating ObjectGroups and Shapes r=sfink
This is a data race, but I don't think it is causing problems at the moment.  It will go away when we get rid of ObjectGroups and is not trivial to fix, hence the decision to suppress it.

Differential Revision: https://phabricator.services.mozilla.com/D81327

UltraBlame original commit: 9872086a16f9be8b4bbcd9ff459acc13af96a4d1