Bug 1645510: Part 0 - Handle first system zone compartment having system principal. r=jandem
Prior to this patch set, XPConnect always created the first compartment in the
system zone with a content principal. The subsequent patches make that
global's creation lazy, which leads us to create the first compartment in the
system zone with the system principal and the NewCompartmentInSystemZone
specifier. In that case, we call `setIsSystemZone()` when we create the zone,
because the compartment has the system principal, and then call it again when
we try to store it in `rt->gc.systemZone`, which leads to a failed assertion.

This patch fixes that.

Differential Revision: https://phabricator.services.mozilla.com/D79718

UltraBlame original commit: 56434e256a52955813614e439b493d551a98a9ee