Bug 1599579 - Part 3: Load toplevel extension frames into a per-extension BrowsingContextGroup, r=kmag
This is done by tracking a specific BrowsingContextGroup ID on the
WebExtensionPolicy in the parent process. Whenever a load is done with that
policy, the browsing context is replaced to ensure it is loaded in the correct
BrowsingContextGroup.

This patch also ensures that extension iframes are always loaded in the same
process as their embedder document, even if the frame was previously remote.

Differential Revision: https://phabricator.services.mozilla.com/D80256

UltraBlame original commit: 73f6c956fc30b46959aa8d5401a8166d61e0fe5d