Bug 1645781 - Part 2: Make nsMIMEInputStream always be serializable. r=baku,necko-reviewers
nsMIMEInputStream was conditionally serializable depending on the
wrapped stream. In general when a stream needs to be sent over IPC, if
it isn't serializable we send it using
InputStreamHelper::SerializeInputStreamAsPipe. There is no reason to
not branch on the wrapped stream when serializing nsMIMEInputStream
and use that, instead of sending the nsMIMEInputStream itself over a
pipe.

Differential Revision: https://phabricator.services.mozilla.com/D80890

UltraBlame original commit: 9df2d1f5a85daf1f1da677f9353d6afd21a3a4e6