Bug 1648881 - Right click submenus opening on the wrong screen r=spohl
Apply the fix for bug 1592416 to submenus.

This adds the NSWindowCollectionBehaviorMoveToActiveSpace behavior to
submenu popups so that they override the "Assign To" space setting and display
on the active space. And, when mutiple displays are in use, recreates the
submenu widget each time it is displayed.

Differential Revision: https://phabricator.services.mozilla.com/D81812

UltraBlame original commit: e7e5921ff49151bf0031ae9f8b73a2b21e717e5c