Bug 1649356: [taskgraph] Don't use `Schema(..., required=True)`; r=ahal
This is the only place where we override the default of `required=False`.
Change it to match the rest of the schemas, so as not to confuse people who
don't happen to notice the final line of the schema.

Differential Revision: https://phabricator.services.mozilla.com/D81657

UltraBlame original commit: 4ffb1ac48cea88dbdcb3c309ec440fabb64596d6