Bug 1645563 - Introduce devtools helper to swallow rejections under a condition r=nchevobbe
We very often have to introduce a pattern in our async methods in order to "swallow"
async errors if a panel was destroyed, if the toolbox was destroyed etc...

This is not an issue in a regular usage of the toolbox, but it can lead to many intermittent failures.
One solution could be to always wait for all requests to be completed before shutting down tests.
Another approach is to have an easy way of swallowing errors when a certain condition is true.

The helper added here will run the provided method in a try catch, and will only bubble up errors if the provided check fails.

Differential Revision: https://phabricator.services.mozilla.com/D81675

UltraBlame original commit: 59f5ba699cc7d0ac97317b09429718624d432024