Bug 1639873 - Add support for test manifest backfilling r=tomprince
We use the symbol of the original task we backfill and suffix the revision of the push
from which we backfill. This makes it a bit more clear as to from which push we requested
the backfill.

We also modify the group symbol (if available) to group backfilled tasks.

***

***

Differential Revision: https://phabricator.services.mozilla.com/D81379

UltraBlame original commit: 3ddf49b60369103142a534092fee1c811f7963bb